Set failOnError on false for JavaDoc
With Java 8 the JavaDoc becomes much more restrictive with the HTML code
and many errors are thrown with no reason. Example: <br/> and </p> are
not accepted. As result of the error the JavaDoc website is not
generated, which is unuseful. This option makes the website be generated
even with errors. BE CAREFUL!