Remove RomanInteger constructor with both parameters
It is not necessary, since while having a pair (Integer, RomanNumeral)
to store, it can be just used one of the two remaining costructors: with
int or with RomanNumeral parameter. In this case the passed value is
going to be converter to the other form. In the case of the
bi-parametric constructor, the two parameters got checked if they match,
so the conversion is done anyway.