Remove RomanConverter accepting Strings as inputs
RomanConverter accepts only RomanNumerals as parameters now.
integerToRomanString() and romanStringToInteger() are now private and
the only public methods available for conversions are
integerToRomanNumeral() and romanNumeralToInteger(), which delegate the
other two methods

This is needed to mantain the correct syntax of the passed roman
numerals to be converted to prevent user errors.

Refactor changes to Test class and test everything.