Create RomanFormatException and add it to RomanNumeral
An extremely simple extension of NumberFormatException which indicates
that the roman syntax is not met during the creation of the RomanNumeral
or during a RomanNumeral setter call. A basic test class for the
RomanFormatException is added.

RomanNumeral is modified to support RomanFormatException:

 - Throws and catches RomanFormatException instead of
NumberFormatException
 - JavaDoc text and links are is refactored
 - "throws" clause is added to constructor and setter
 - Test class is refactored

RomanConverter.integerToRomanNumeral(int arabic) now has "throws"
RomanFormatException clause.